requirements.txt
Pillow==10.2.0
opencv-python==4.9.0.80
ffmpeg-python==0.2.0
numpy==1.26.4 